fix color types and hooks
