+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from "react";
 import { Geist, GeistMono } from "~/components/font";
-import Priest from "~/components/ui/priest";
+import Layout from "~/components/ui/layout";
 import clsx from "clsx";
 import "./style.css";
 
@@ -9,7 +9,7 @@ export default function RootLayout(props: Readonly<{ children: ReactNode }>) {
     <html lang="en" className={clsx(Geist.variable, GeistMono.variable)}>
       <body>
         <div id="root" role="none">
-          <Priest>{props.children}</Priest>
+          <Layout>{props.children}</Layout>
           <div id="portal" role="none"></div>
         </div>
       </body>
